Fix for the issue of replacement video not loading.
The video playback issue may not necessarily be due to the moov atom being at the end of the file; it could also be related to special packaging issues. If every video is read with ffprobe, the loading speed will be particularly slow. Now, if there is a playback block error, directly run ffmpeg with faststart to replace the original file.